Java 'script service': fix executing DynamicCommands
To execute a DynamicCommand, the PluginService needs to be able to find
the PluginInfo since it is asked to find the PluginInfo for the current
Command when the context gets set.

In the same run, we need to make sure that the Plugin annotation is known
to the PluginInfo. Otherwise the ModuleService cannot find the
"initializer" setting, for example.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>