Revert "Use more descriptive label for legacy plugins"
The idea was to make the command finder more useful, by showing not only
the last leaf of the menu path.

It was also meant to help the situation when we search for Image>Zoom>In
and type 'zoom'. But we will solve that issue in the command finder
directly.

This reverts commit 7d086d2232981b263b353dc10f41a22a578d4e02.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>