Main: split the logic into its own launch method
It is nice to be able to keep a handle on the newly created ImageJ
context after it is launched. But unfortunately, a Java main method must
return void. So we split the logic into another static method called
"launch" which returns the context, and then call launch from main.

This allows external code to now call Main.launch(args) instead of
Main.main(args) to retain a handle on the ImageJ context, with which
they can perform subsequent operations such as running commands.