Launcher: Add the --debugger=<port> option to make Eclipse debugging easy
In Eclipse, you can attach to a running JVM if it has been called with
JDWP enabled. The newly introduced --debugger=<port> option lets you
do that.

When ImageJ was started that way, you can do the following in Eclipse:
Run>Debug Configurations>Remote Java Application, and in the context
menu create a new configuration. All you need to do now is to select
the ij-app project and the appropriate port (which defaults to 8000,
so you might want to use --debugger=8000 to start ImageJ accordingly).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>