DefaultPluginFinder: fail gracefully when plugin.type() fails
When dependencies are not installed or not at an appropriate
version, it is possible that plugin.type() throws an exception such
as ClassNotFoundException. This is unfortunate, but by no means a
reason for us to give up on all other plugins, or for that matter,
on creating a valid ImageJ context.

Following Postel's law, we therefore fail gracefully by simply ignoring
the plugin whose type could not be instantiated.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>