Updater: ignore dependencies that became obsolete
The issue came up with Bio-Formats' rename of loci-common to loci-legacy.
That change was performed on the ImageJ update site, so the dependencies
on the Fiji update site (or for that matter, any other update site
offering files depending on loci-common) were not updated.

Heeding Postel's law, we simply ignore such dependencies for the time
being now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>