PluginService: improve getPluginsOfClass methods
The getPluginsOfClass(Class, *) methods now work a bit differently, to
ensure that the resultant PluginInfo objects have not only a matching
class *name*, but also a matching *Class* object. To do this, they first
ensure that the class names match, then load the matching Classes as
needed and exclude the ones whose Class objects differ from the given
one. This is necessary because it is possible to have more than one
Class object for the same fully qualified class name, if the classes
were loaded using different ClassLoaders.

There is also a new method getPluginsOfClass(String, Class) for limiting
the search to a particular plugin type. (This method is analogous to the
getPluginsOfClass(Class, Class) method, which already existed.)