Initial support of results table synchronization
With this commit ResultsTables are translated bidirectionally across
the legacy layer.

This is an initial implementation that needs to be refactored out into
an external class used by LegacyCommand.

Currently the ResultsTable of the most recently active TableDisplay is
the one given to IJ1.

There are known bugs with fixes to follow.