Be smarter when autofilling single Display inputs
This change utilizes the improved "active display" behavior that
supports querying of the most recent active display by display type.

Note that it will not autofill display parameters if a command declares
multiple Displays of different types; the preprocessor will fail to
detect a single unresolved Display input, and pass it over, forcing the
user to fill in the values manually. But the behavior should not be any
worse than it was before this update.