Launcher (Windows): only open console by default if the name is debug.exe
We had some cute code that would automatically open the console if the
executable was named ImageJ.exe. The idea was actually to open the
console if it was *not* named ImageJ.exe, but even that was misguided.

Instead, only launch the console if the name is 'debug.exe' which we can
tell users to do easily (it is much easier to rename a file on Windows
than to start it with command-line options).

Noticed by Wayne Rasband.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>