Launcher: heed the JRE setting in legacy ImageJ.cfg
The second line of the (not very well-documented) ImageJ.cfg format
is supposed to specify the path to javaw.exe.

If it points to java.exe instead, the launcher is expected to open
a console.

This entails basically a rewrite of the parse_legacy_config() function
since we used to only ever look at the third line. As a bonus, we
handle the case correctly now where ImageJ options were specified
after the main class.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>