Skip legacy tests in headless mode
The tests were not executing properly in headless mode already. This
change merely formalizes that limitation, so that rather than the tests
erroring out, they are simply skipped. This allows Jenkins to continue
to build ImageJ successfully.

The downside is that we still have no CI mechanism for verifying that
the legacy layer is working properly. For now, we must take care to
occasionally run the unit tests ourselves on a system with a display.