LogService: add the concept of a logging level
This organizes the five log types into a hierarchy, and provides an API
for accessing and mutating that log level.

It also adds support for log configuration via a Java system property,
imagej.log.level, in the default log service (StderrLogService). And if
the system property is not set, the DEBUG environment variable is used
like before.