UIService: add show(Object) methods
This makes it dirty easy to display any object using ImageJ.

It works by first asking the display service for an appropriate display
for the given object. It then shows that display using an appropriate
viewer for the default UI. Or you can tell a specific UI to show the
object itself, if you prefer.