Finish eliminating obsolete code
The new ThreadGroup tracking code used by LegacyOutputTracker
can be used to eliminate the old special case code we had for
carefully tracking who initialized an image close (IJ1 or IJ2).
Now the legacy injected methods can always determine whether to
record an ImagePlus in the tracker or not depending upon whether
there is a LegacyCommand ThreadGroup active or not.