Harmonize ImagePlus instances when switching to the legacy mode
When an image was opened in non-legacy mode, and no legacy command
called on it, switching to legacy mode would not switch that image to
legacy mode. Fix this by doing what the LegacyCommand would have done
in that case.

Note: this is a strong hint that the code executed by LegacyCommand
does not belong there in the first place, since it has nothing to do
primarily with running a legacy command, but more to do with mapping
images with their legacy counterparts.

Noticed by Barry DeZonia and Curtis Rueden.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>