Populate the menus with Lookup Tables
The DefaultLutService now populates the menus with color table
menu entries that apply a given color table to the current image.
The tables to use are discovered in the "luts" subdirectory off of
the base directory.

Unlike IJ1 we allow luts to be organized in subdirectories that
each get their own submenu.