ModuleRunner: inject context into the module
Not all modules are contextual. But it is important for the context to get
set when the module *is* contextual, so that any declared EventHandler
methods are registered with the event service properly.

This change fixes the ShadowsDemo to behave properly again, after being
broken in commit f69f6fd081c1a82f110cb8b9f54e68bd98489be9.