UserInterface: show new displays using the EDT
If we don't do this, at least with Swing, JHotDrawImageCanvas throws an
exception in its sync(boolean) method. The sanity check there is not
spurious; the syncing really doesn't work when not run from the EDT.

It is possible that in the future, there will be a UI that is
thread-safe, such that new displays can be shown off the EDT. If/when
that occurs, we can refactor the code to do the right thing. But for
now, this solution suffices.