Add the 'populate-app' profile
Instead of making a zip and unzipping it to populate an ImageJ.app/
directory, it is better to use the assembly plugin's option to output
to a directory directly.

Suggested by Curtis Rueden (a long time ago, actually).

However, we cannot use it in bin/populate-app.sh, unfortunately:
there is no way to force existing files to be overwritten with the
assembly plugin:

	http://jira.codehaus.org/browse/MASSEMBLY-478

Therefore, we are unable to discern between obsolete and up-to-date files
in the ImageJ.app/ directory structure after running

	mvn -Ppopulate-app

Since we still have to rename, merge or split artifacts, we need to know
about obsolete files, though.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>