bin/populate-app: use -Ppopulate-app upon initial use
Unfortunately, we do not have a way to identify obsolete files when
using the 'dir' format with the maven-assembly-plugin. But we can
use it to initialize ImageJ.app (because there cannot be any obsolete
files then, and it is slightly faster to avoid building and unpacking
a .zip file).

Since this commit changes the indentation of bin/populate-app.sh, it is
best viewed with the '-w' diff option of Git.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>