AboutImageJ: use copyright range to current year
Elsewhere in the project (i.e., in copyright headers and LICENSE.txt
files), we use a date range rather than explicitly listing every year.
According to the GNU GPL How-To guide:

    For software with several releases over multiple years, it's okay to
    use a range ("2008-2010") instead of listing individual years
    ("2008, 2009, 2010") if and only if every year in the range,
    inclusive, really is a "copyrightable" year that would be listed
    individually; and you make an explicit statement in your
    documentation about this usage.

(See http://www.gnu.org/licenses/gpl-howto.html for the full text.)

In the case of ImageJ, every year in the range is, and will continue to
be, a "copyrightable" year. The only issue is whether we have made "an
explicit statement" in our documentation. Unfortunately, the guide does
not explain any further what such a statement might look like. There are
thousands of projects on the web that use date ranges and it is
unambiguous what they mean, so I think we are safe with this practice.

Lastly, continuing to enumerate all the years individually will run out
of real estate on the image within a few years' time, whereas using a
date range will scale well into the foreseeable future.