SwingDialog: remove buttons if dialog is non-modal
This brings the behavior of interactive commands in line with how ImageJ
1.x does things. It doesn't make sense for a non-modal dialog to have OK
and Cancel buttons, since there is nothing that can consume the result.

Note that it is still possible to have buttons in a non-modal dialog
by using the imagej.widget.Button class; see
imagej.core.commands.debug.ButtonDemo for an illustration.