ClearRecent: services are not outputs
I think I see what I was thinking here: when performing actions that
manipulate the state of the service, it could be argued that that
service is an "output" similar to how ImageDisplays are ItemIO.BOTH when
the command manipulates them. But for now, service outputs are
unsupported anyway, and there is really no need to chain them along in a
workflow pipeline anyway, since they always get populated properly at
each step of the way.