Use checked exceptions for reflective method calls
In particular, the initialize() and callback() methods in the module
framework really benefit from the new checked exception,
MethodCallException.

It is now the responsibility of downstream code to handle this
exception. Previously, it was (too) easy to ignore when the reflective
method execution failed, which resulted in some exceptions getting eaten
further down the chain. For example, a NumberWidget for a double input
parameter with a scroll bar or slider widget style does not work, but
the exception indicating so was previously silently ignored.