Launcher: add the '--set-icon' command for Windows
On Windows, the icon of .exe files can be set by updating the resources
included in the .exe file using functions of the Win32 API.

Since it is so hard to get by easy-to-run programs to do that from
scripts, and since we want to be able to rebrand the ImageJ launcher
easily both for Fiji and for ImageJ 1.x, we simply add a subcommand of
the launcher that can set the icon (but this only works on Windows, as
it requires Win32 API calls).

Therefore this option is meant primarily to be run by Jenkins on a
Windows node to copy-edit the launcher for Fiji and ImageJ 1.x.

Unfortunately, there seems to be no error message when trying to update
the icon of the running program, even if the icon is not updated. Caveat
emptor.

This commit adds a BSD-licensed reimplementation of a small part of the
code of WinRun4J. It was written from scratch after learning from
WinRun4J's source code, the MSDN documentation and reverse engineering
the output of windres.exe.

However, it is this software developer's conviction that -- modulo
naming -- there is only one way to implement this functionality, so it
is not copyright'able at all.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>