InteractiveCommand: rename field "inputs" to "listenerNames"
One of the base classes of InteractiveCommand already defines an "inputs"
field. In the debugger this was confusing. Rename this version to
"listenerNames" because that is closer to the truth.

While at it I changed updateInputs() to updateListeners(). This event
handler did not appear to be referenced elsewhere and thus this rename
should be safe.