Split I/O service and commands
The I/O service continues to live in ij-io, while the I/O commands are
now part of ij-commands.

There is still some cleanup to do in the NewImage and AWTDropListener
classes to migrate logic into services more properly. Right now, the
ij-ui-common-awt project has a dependency on ij-commands (so that
AWTDropListener can invoke the NewImage and OpenImage commands), which
is suboptimal. Fixing this issue may go hand in hand with ticket #860
anyway, so for now we wait.