WIP: Can we have an ImageJ2 command that is *also* an ImageJ1 PlugIn?
Why not? As long as there is a plugins.config for it (or the class has
underscores)... but Fiji may need some massaging. Because ideally, an
ImageJ2 context would not be created on the fly each time, but rather
obtained from ImageJ1. With some bytecode manipulation, we can inject
ImageJ1 with an associated ImageJ2 context, and access it via a static
helper method which uses reflection. (Are we doing this already?)

The end result is that you can write a class which extends
LegacyCompatibleCommand and use all the goodness of ImageJ2, while
operating on ImageJ1 data structures like ImagePlus, ImageStack and
ImageProcessor!