Add a convenience method for the Bug Submitter
We will most likely make the Bug Submitter much more modular (so
that we can have a Trac backend, for example) and move it from Fiji
to ImageJ2, but for the time being, it is better to use compile-time
binding between Fiji's Bug Submitter and the updater.

While at it, shorten the checksums to 8 characters so that the bug reports
are not that confusingly long.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>