Add imagej.ui.dnd package and event hierarchy
The imagej.ui.dnd package defines a new DragAndDropService which
provides a UI-agnostic way of handling drag and drop operations.

Similar to keyboard and mouse events, the imagej.ui.dnd.event package
provides a UI-agnostic set of events for receiving notification of
drag-and-drop events, which the DragAndDropService can utilize.

The imagej.ui.dnd package also defines a new DragAndDropHandler plugin
type to make drag-and-drop behavior extensible and customizable.