AbstractData: remove explicit isDiscrete() checks
It is the responsibility of downstream implementations (e.g.,
AbstractOverlay) to throw UnsupportedOperationException, or not,
depending on its needs when calling into the methods if Interval. The
AbstractData implementation of getExtents() and getDims() will just try
to ask the Interval for that information regardless of the value of
isDiscrete(), because in some cases a reasonable approximation can be
provided.

Specifically, Overlays can return the exterior integer bounds for their
actual region; they have actually been doing this for a long time. So
asking an Overlay for its Extents will return such information, which
can be useful in some circumstances such as cropping.

See ticket #1335 for related details.