Updater (cmd-line): only call System.exit() when run stand-alone
As we want to reuse the CommandLine updater in the unit tests,
we cannot allow it to end the current process except when run as a
stand-alone main class.

We do this by teaching the die() method to exit only when the class
was called from the main() method.

To help the compiler detect unreachable code correctly, we do
not simply throw an exception in embedded mode, but return the
RuntimeException that has to be thrown by the caller.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>