Rename AppService hierarchy to AppEventService
This is in preparation for a new AppService to be introduced into the
SciJava Common framework itself, for handling things like title and
version. The ImageJ AppEventService is intended for handling the
UI-level operations of "show an about dialog box", "show the preferences
dialog box" and "quit the program with user interaction".