Use PluginService#createInstance where feasible
This method is better than PluginInfo#createInstance(), because it
automatically injects the context and priority, catches exceptions, and
logs them to the LogService. We were doing all of this manually in a few
places, which needlessly repeated code, and increased the probability of
bugs related to context and/or priority injection (or lack thereof).