SystemInformation: report info on all apps
The AppService reports information on all registered Apps, so why not
have System Information report all that information, too, rather than
only that of the ImageJ app? For example, if SCIFIO is present (which it
will be if ij-io is on the classpath), System Information will now
include information about it as well, such as its version number.