ImageJ: extend new SciJava Common Gateway class
This bumps the scijava-common dependency to 1.2.0 to gain access to the
new Gateway class hierarchy. It slightly simplifies the ImageJ gateway
implementation, because the get method implementations are now provided
in the abstract superclass.

The ImageJ gateway now also automatically inherits the null checks, and
so its accessors will now throw NoSuchServiceException when services are
missing (rather than simply returning null).