Move all LUTs into ij-data.jar
We provide a set of default lookup tables. We want users to be
able to override them by putting their own version into the luts/
directory. If they would have to override ImageJ's own files, the
updater would always mark them as locally modified. That is why we
added the option to put default lookup tables into .jar files so
that the originals do not have to be overwritten by the user. But
that means that we should also use that option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>