Fix finding .lut files in .jar files
We really need to look in all .jar files.

Besides, there is no use in listing all files in a directory and its
subdirectories, then throwing away the information which base directory
we were looking into, and then trying to reconstruct said base directory
from the URLs.

Instead, we need to discover LUTs as String/URL combinations, the String
being the menu path and the URL being the location where to find the
actual lookup table. Hard-coding the expectation that the base
directory's suffix is '/luts/' and the additional assumption that no
submenu in Image>Lookup Tables can ever be named 'luts' is not something
we are going to do.

This breaks the public API of the LutFinder in a rather blatant way, but
since we are still in beta phase and since it is not a cross-project API
for the moment (no collaborator uses the LutFinder directly), we can
still fix that without having to deprecate things.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>