Fix intermittent test failures
The testReCecksumming test succeeded only by mistake, and only most
of the time: the checksummer would see that there is a .checksums
file with cached checksums, see that the timestamps matched, and
would fail to re-checksum the files.

Test properly that files are re-checksummed, and make sure that the
checksummer has no .checksums file for a reason to procrastinate.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>