Another microscoff time tax
In instances where still-locked files are about to become unused,
Windows makes this a problem: it refuses to delete them. Typically you
work around this by calling File#deleteOnExit. But if you have to delete
directories recursively, no such easy way is available. So we have to
make it available.

I wish we would get something back in return for having to spend so much
time to work around design flaws in Windows for a change.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>