Add new MIME type helper class
I was reluctant to do it, but the MIME types are complex enough (with
the parameter list of key/value pairs) that it seems necessary.

AWT already has a MIME type helper class,
java.awt.datatransfer.MimeType, but: A) it is part of AWT, which is not
available in all Java environments, particularly Android; and B) it is
not a public class anyway.

Pivot also has a MIME type helper class, org.apache.pivot.util.MIMEType,
but we do not want to add a Pivot dependency directly to ij-ui.

So unfortunately we reinvent the wheel again here.