Make drag-and-drop handlers compatible with nulls
AWT uses Transferable objects that internally have a list of
DataFlavors. Each DataFlavor has a MIME type and associated Java class.
You can ask a Transferable to give back its data in any supported
DataFlavor. Unfortunately, if you do this too early in the process (such
as during the DropTargetListener#dragEnter event of a drag-and-drop
action), an InvalidDnDOperationException is thrown because AWT does not
consider the system to be in an appropriate state to extract the data.

As a consequence, the drag-and-drop handlers cannot obtain the data
currently being dragged (e.g., not even the filenames of dragged files).
Hence, the isCompatible(DragAndDropData) methods cannot inspect the
dragged data in any more detail than the MIME type and Java class.

To work around this, the AWTDragAndDropData class's getData(MIMEType)
implementation now catches InvalidDnDOperationException and returns
null, and the UI-agnostic ImageJ drag-and-drop handlers have been
changed to trivially accept null data. The handlers' drop
implementations also trivially return true, although in practice at that
point (DropTargetListener#drop event) they should be able to obtain the
actual non-null data via the Transferable.