DragAndDropService: show only if drag is supported
This is how Barry had done it before, but I thought I'd be clever and
display more information about the drag-and-drop handler chosen.
However, in practice, it's pretty much always the ListDragAndDropHandler
(at the top level at least). So the message is pretty useless, and it's
not friendly to the user to show a fully qualified Java class name.

So we revert to the simple "Drag and Drop" message instead, for now...