ImageFileDragAndDropHandler: make low priority
Otherwise, with the current SNAPSHOT version of SCIFIO, the PGM reader
claims it can handle LUT files (and then subsequently barfs when trying
to actually read the file), so the LUTFileDragAndDropHandler never gets
its chance to shine.

Hopefully this will be resolved with the SCIFIO refactoring, but for
now, it's safer to make the ImageFileDragAndDropHandler take a back seat
to some of the other more conservative drag-and-drop handlers.