Broadcast DnD events from Swing ImageDisplays
Specifically, we listen for DnD events on the JHotDrawImageCanvas, and
rebroadcast them to the UI-agnostic drag-and-drop framework using an
AWTDropTargetEventDispatcher.

It is now possible to drag stuff onto the main part of an image display
and have something happen. In particular, you can drag a LUT onto an
image and it will change the image to use that LUT. Groovy.