SwitchToModernMode: initialize Context if necessary
We want this plugin to work even if there was no SciJava Context
yet... i.e. if this plugin is run from ImageJ 1.x without ImageJ2
running before that, we still want things to work nicely.

This will be needed for the upcoming Help>Switch To Modern Mode
support in Fiji.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>