DefaultLegacyService#dispose: adjust IJ._legacyService, too
When we dispose of a LegacyService, we need to make sure that all
references we know about are released. For convenience, we introduced a
private static field in the ij.IJ class (and a public static read-only
accessor) that needs to be re-set to a dummy legacy service (our
contract is that it cannot be null).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>