LegacyImageMap: ensure that ImageJ 1.x has been patched before use
When running the LegacyImageMapTest before the, say,
OverlayHarmonizerTest, the ImageJ 1.x classes were used but not patched.
For that reason, the subsequent tests failed just because previous tests
did not require the LegacyService.

This pointed to a bigger issue: whenever the LegacyImageMap -- or for
that matter, any class using ImageJ 1.x classes -- is loaded before the
LegacyInjector could do its thing, subsequent initialization of the
LegacyService will fail badly.

At least with LegacyImageMap we can do something about it: call a static
method (whose return value we actually do not care about) of the
DefaultLegacyService to ensure that the static block initializing the
LegacyInjector has been executed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>