Start up legacy service in legacy mode
While the LegacyService is being initialized, we cannot rely on any
ImageJ2 services yet. So the best idea is to say that we're in legacy
mode. If an ImageJ 1.x instance is currently running, our best bet is
to stay in legacy mode. Otherwise, let's switch to modern mode just
after initialization is complete -- toggleLegacyMode() knows that it
should not try to show anything if no UI is available.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>