Add a workaround for the "white display" problem
This manifests, e.g., when running the load-and-display-dataset ImageJ
tutorial, among other places. The workaround just updates the display
again after showing it, while still on the event dispatch thread.

Of course, all the DisplayViewer logic is currently convoluted and
overcomplicated; we have plans to simplify it to a large extent on the
display-refactor branch. But in the short term, this fix does the trick.