BrightnessContrast: add min/max calculation FIXME
Right now, B&C unnecessarily recomputes the min & max of the entire
Dataset, even if the metadata is already present in the ImgPlus. And if
it is not present, there should be a central way to populate it lazily
rather than the B&C command being responsible for doing it.