Typed, HandlerService: API tweaks
The Typed#supports method is now actually typed on the generic
parameter, rather than accepting any Object. This made things much
more elegant in several places.

The HandlerService#getHandler method is now typed directly with its DT
and PT parameters, rather than as extension points. The extensions were
unnecessary and introduced additional complexity. The return type of
getHandler was also not guaranteed to be a PT at compile time, which was
extremely unintuitive and silly.