AbstractTypedService: remove abstract ctor layer
Constructors in Java have some properties which make them a bit unwieldy
with inheritance: in particular, non-default constructors are not
automatically available to subclasses. Further, including a non-default
constructor in an abstract superclass forces a compile-time requirement
on subclasses to provide at least one constructor of their own.

To avoid these difficulties, the AbstractTypedService's "convenience
constructor" was actually not much of a convenience. So it has been
removed in favor of requiring concrete subclasses to implement the
getType() and getPluginType() methods directly. The amount of
boilerplate ends up being similar, but without introducing any
non-default constructor messiness.