HandlerService: extend SingletonService
This change alters HandlerService to extend the new SingletonService in
addition to TypedService. It also introduces a corresponding dedicated
HanderPlugin for clarity and consistency, which is both a
SingletonPlugin and a TypedPlugin.

The DragAndDropService (which was already a HandlerService) has also
been updated accordingly to reflect the updated HandlerService API.