AbstractTypedPlugin: remove ctor
The reasoning is similar to 3ad14cf87cb09f88807df07b5738b9e20a1a2938:

Constructors in Java have some properties which make them a bit unwieldy
with inheritance: in particular, non-default constructors are not
automatically available to subclasses. Further, including a non-default
constructor in an abstract superclass forces a compile-time requirement
on subclasses to provide at least one constructor of their own.

To avoid these difficulties, the AbstractTypedPlugin's "convenience
constructor" was actually not much of a convenience. So it has been
removed in favor of requiring concrete subclasses to implement the
getType() method directly. The amount of boilerplate ends up being
similar, but without introducing any non-default constructor messiness.