WebDAV uploader: support spaces in file names
When uploading files with special characters in their file name, we need
to be extra careful. Spaces need to be passed as %20, and pretty much any
non-alphanumerical character has a special meaning.

Unfortunately, URLEncoder encodes for form values, and that means that
spaces are converted into plusses, and slashes are handled, too. So we
have to undo a little bit there.

Reported by Curtis Rueden as http://trac.imagej.net/ticket/1867.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>