Updater (Swing): allow Unicode passwords
Internally, Java still uses ISO-8859-1 which -- if you think long and
hard about it -- might have been slightly shortsighted, but at least
one can understand that if English was good enough for Jesus, it is
good enough for any programming language. The only funny thing there
is that even variable names in Java can contain umlauts. But I digress.

As mentioned above, internally, Java still uses ISO-8859-1 and while
the authenticator purports to work on char arrays for passwords
instead of byte arrays, internally it simply casts chars to bytes,
in effect handling char arrays as if they were raw bytes:

http://hg.openjdk.java.net/jdk7/jdk7/jdk/file/96c75aec5545/src/share/classes/sun/net/www/protocol/http/BasicAuthentication.java

So while it would be awesome to work with properly encoded Strings
all the time, we cannot... Let's work around this issue by treating
the password as the UTF-8-encoded String it should have been.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>