Rename DefaultModule* to DefaultMutableModule*
The reason is that these default implementations implement not only the
Module* interfaces, but the MutableModule* ones. It is misleading to
call them DefaultModule* when in fact they are more than that.

This change also splits the old DefaultModuleInfo into two parts: an
AbstractModuleInfo base class, and a DefaultMutableModuleInfo default
implementation. This brings it in line with the Module and ModuleItem
interfaces, which also have both an abstract base class and a default
mutable implementation.