ModuleInfo: add type-safe item accessors
It is now possible to ask for an input or output item by both name and
type. The behavior is the same as asking only by name, except that if
the item of that name exists but is the wrong type, the method throws an
IllegalArgumentException. In this way, there is now a compiler-friendly
way to ask for inputs and outputs without explicit casting.