MutableModule: explicitly promise a mutable info
It was always the case that our existing MutableModule implementation
(DefaultMutableModule) has a MutableModuleInfo. But it was not
previously guaranteed at the interface level, only in the
DefaultMutableModule implementation itself. It is nicer to guarantee it
at the interface level to avoid downstream code needing to cast and/or
rely on something being an instance of the default implementation.