Remove crappy and confusing shell scripts
I originally added these scripts before we had integrated the ImageJ
launcher. Then I left them there because (I thought) they might serve as
a convenient alternative to launch ImageJ in a "quick and dirty" way if
the launcher didn't work for some reason. But that is really a bad idea,
because the way they construct the classpath differs, and will not work
correctly in many cases due to ImageJ's usage of custom class loaders.

So let's just get rid of them, to avoid problems and confusion!