Updater: support (un)shadowing files
We call an upload 'shadowing' when there is a lower-ranked update
site which also serves that file (because the higher-ranked site will
shadow the lower-ranked site's version).

Likewise, removing a file from a higher-ranked update site which is
served by a lower-ranking update is called 'unshadowing'.

This commit adds support for that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>