Updater: prepare the FileObject class for 'unshadowing'
When removing (i.e. marking obsolete) a file on a give update site,
this 'unshadows' a lower-ranked update site's version of the same file.

This commit prepares the FileObject class for the situation when the
user marked a file to be unshadowed, but then made up her mind again,
and reverted that decision.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>