Updater: move the UpdateSite class into its own file
While at it, enforce the use of getters and setters (as seen with
the URL and the upload directory, it is a good idea to have setters
to be able to enforce certain formats.

Since this is an API-changing commit, it requires fixes in all uploaders,
and in Fiji's Fix_Updater.js, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>