MediaWikiClient: add support for creating an Fiji Wiki account
This is required for better support of the personal update sites:
users should be able to initialize their own update sites from within
the updater's GUI: some people might not have a Fiji Wiki account
(because so far, we recommended creating such an account only for
people who actually wanted to be helpful and contribute something on
the Fiji Wiki).

There is a rudimentary security-by-obscurity measure in place to help
prevent flooding the Wiki with spammers' annoying and irritating
'content', but the idea is really to be light-weight on that side:
let's try harder to prevent such things only when we have to, and go
back to writing code for scientific image analysis in the meantime.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>