Updater (Swing) Sites Dialog: change the real upload information
When an update site listed in http://fiji.sc/List_of_update_sites was
active, by mistake, we used a copy of the locally-listed update site
in the dialog. As a consequence, only that copy's upload information
was updated, not the information that would get persisted in db.xml.gz
(and hence the upload information entered by the user would be simply
ignored).

Fix this by using the local copy, filling in the information from
the Fiji page, instead of the other way round.

Reported by Ignacio Arganda-Carreras.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>