Added Plane/Dataset toggle to BrightnessContrast
The BrightnessConstrast plugin now has a toggle to pick if the min/max
values are computed over the entire dataset or just the current plane of
the underlying view.

Default value is to use the current plane.

This allows brightness/contrast to be run on incredibly large datasets.

This should be a part of the PlanarCommand API, which requires
restructuring of commands. Thus this change serves as a proof-of-concept.